Add index.html template and update application properties
The new index.html template has been added to the resources/templates directory. This is the main page for the 'Working Place Booking' application. It includes a navbar, a showcase section, as well as a footer. Application properties now include thymeleaf prefix and suffix for viewing HTML as well as a server servlet context path for dynamic website adaptation.