Configure H2 database and downgrade Spring Boot version
H2 database has been set up and configured in the properties file to use JDBC. Configurations for console path, tracing, remote access and more are also established in the properties file. Additionally, the Spring Boot version in the pom.xml has been downgraded from 3.2.1 to 3.1.7. This also includes adding the database directory to the .gitignore file.