add the role functionality, add, delete ...
Service classes are moved to corresponding sub-packages for a more structured representation. CRUD operations for User and Role are enhanced to handle both creation and updates. Actions on users and roles are reflected in corresponding templates, with URL mappings correctly pointing to respective CRUD operations. Additionally, the user's password field is updated to display as plain text.