add Cancel button in user formula template
Adjustments have been made to the userFormula.html template. The label for the password field has been replaced with a hardcoded "Password" string. Additionally, the submit button has been wrapped in a div and a cancel button has been added. This provides users with a clearer form and an option to cancel the action.