Refactor UserServiceImpl and update UI layout
The UserServiceImpl class was refactored to simplify the update and delete methods. All truth checking is now moved inside the update method, thus returning void. Addition of passwordEncoder into UserServiceImpl is planned for hashing the user's password. Adjustments were made to the bookingDesk.html layout to centralize content. Additionally, a comment explanation was added in UserController for id-related user operations.